Implemented mask_motion_scale into Motion Model Settings. Refactored a lot of code, moved extra/experimental/deprecated nodes into separate files,
