+++ animatediff/motion_module_ad.py
@@ -5,9 +5,11 @@ import torch
 from einops import rearrange, repeat
 from torch import Tensor, nn
 
+from comfy.utils import repeat_to_batch_size
 from comfy.ldm.modules.attention import FeedForward
+from controlnet import broadcast_image_to
 from .motion_lora import MotionLoRAInfo
-from .motion_utils import GenericMotionWrapper, GroupNormAD, InjectorVersion, BlockType, CrossAttentionMM
+from .motion_utils import GenericMotionWrapper, GroupNormAD, InjectorVersion, BlockType, CrossAttentionMM, TemporalTransformerGeneric, prepare_mask_batch
 
 
 def zero_module(module):
@@ -58,14 +60,14 @@ class AnimDiffMotionWrapper(GenericMotionWrapper):
         # but only after implementing a fix for lowvram loading
         return self.loras is not None
     
-    def set_video_length(self, video_length: int):
+    def set_video_length(self, video_length: int, full_length: int):
         self.AD_video_length = video_length
         for block in self.down_blocks:
-            block.set_video_length(video_length)
+            block.set_video_length(video_length, full_length)
         for block in self.up_blocks:
-            block.set_video_length(video_length)
+            block.set_video_length(video_length, full_length)
         if self.mid_block is not None:
-            self.mid_block.set_video_length(video_length)
+            self.mid_block.set_video_length(video_length, full_length)
     
     def set_scale_multiplier(self, multiplier: Union[float, None]):
         for block in self.down_blocks:
@@ -75,6 +77,14 @@ class AnimDiffMotionWrapper(GenericMotionWrapper):
         if self.mid_block is not None:
             self.mid_block.set_scale_multiplier(multiplier)
 
+    def set_masks(self, masks: Tensor, min_val: float, max_val: float):
+        for block in self.down_blocks:
+            block.set_masks(masks, min_val, max_val)
+        for block in self.up_blocks:
+            block.set_masks(masks, min_val, max_val)
+        if self.mid_block is not None:
+            self.mid_block.set_masks(masks, min_val, max_val)
+
     def set_sub_idxs(self, sub_idxs: list[int]):
         for block in self.down_blocks:
             block.set_sub_idxs(sub_idxs)
@@ -82,6 +92,14 @@ class AnimDiffMotionWrapper(GenericMotionWrapper):
             block.set_sub_idxs(sub_idxs)
         if self.mid_block is not None:
             self.mid_block.set_sub_idxs(sub_idxs)
+    
+    def reset_temp_vars(self):
+        for block in self.down_blocks:
+            block.reset_temp_vars()
+        for block in self.up_blocks:
+            block.reset_temp_vars()
+        if self.mid_block is not None:
+            self.mid_block.reset_temp_vars()
 
 
 class MotionModule(nn.Module):
@@ -102,17 +120,25 @@ class MotionModule(nn.Module):
             if block_type == BlockType.UP: 
                 self.motion_modules.append(get_motion_module(in_channels, temporal_position_encoding_max_len))
     
-    def set_video_length(self, video_length: int):
+    def set_video_length(self, video_length: int, full_length: int):
         for motion_module in self.motion_modules:
-            motion_module.set_video_length(video_length)
+            motion_module.set_video_length(video_length, full_length)
     
     def set_scale_multiplier(self, multiplier: Union[float, None]):
         for motion_module in self.motion_modules:
             motion_module.set_scale_multiplier(multiplier)
     
+    def set_masks(self, masks: Tensor, min_val: float, max_val: float):
+        for motion_module in self.motion_modules:
+            motion_module.set_masks(masks, min_val, max_val)
+    
     def set_sub_idxs(self, sub_idxs: list[int]):
         for motion_module in self.motion_modules:
             motion_module.set_sub_idxs(sub_idxs)
+
+    def reset_temp_vars(self):
+        for motion_module in self.motion_modules:
+            motion_module.reset_temp_vars()
 
 
 def get_motion_module(in_channels, temporal_position_encoding_max_len):
@@ -152,20 +178,32 @@ class VanillaTemporalModule(nn.Module):
                 self.temporal_transformer.proj_out
             )
 
-    def set_video_length(self, video_length: int):
-        self.temporal_transformer.set_video_length(video_length)
+    def set_video_length(self, video_length: int, full_length: int):
+        self.temporal_transformer.set_video_length(video_length, full_length)
     
     def set_scale_multiplier(self, multiplier: Union[float, None]):
         self.temporal_transformer.set_scale_multiplier(multiplier)
+
+    def set_masks(self, masks: Tensor, min_val: float, max_val: float):
+        self.temporal_transformer.set_masks(masks, min_val, max_val)
     
     def set_sub_idxs(self, sub_idxs: list[int]):
         self.temporal_transformer.set_sub_idxs(sub_idxs)
 
+    def reset_temp_vars(self):
+        self.temporal_transformer.reset_temp_vars()
+
     def forward(self, input_tensor, encoder_hidden_states=None, attention_mask=None):
         return self.temporal_transformer(input_tensor, encoder_hidden_states, attention_mask)
+        #portion = output_tensor.shape[2] // 4 + output_tensor.shape[2] // 2
+        portion = output_tensor.shape[2] // 2
+        ad_effect = 0.7
+        #output_tensor[:,:,portion:] = input_tensor[:,:,portion:] * (1-ad_effect) + output_tensor[:,:,portion:] * ad_effect
+        #output_tensor[:,:,portion:] = input_tensor[:,:,portion:] #* 0.5
+        return output_tensor
 
 
-class TemporalTransformer3DModel(nn.Module):
+class TemporalTransformer3DModel(nn.Module, TemporalTransformerGeneric):
     def __init__(
         self,
         in_channels,
@@ -187,6 +225,7 @@ class TemporalTransformer3DModel(nn.Module):
         temporal_position_encoding_max_len=24,
     ):
         super().__init__()
+        super().temporal_transformer_init(default_length=16)
 
         inner_dim = num_attention_heads * attention_head_dim
 
@@ -216,27 +255,35 @@ class TemporalTransformer3DModel(nn.Module):
             ]
         )
         self.proj_out = nn.Linear(inner_dim, in_channels)
-        self.video_length = 16
 
-    def set_video_length(self, video_length: int):
+    def set_video_length(self, video_length: int, full_length: int):
         self.video_length = video_length
+        self.full_length = full_length
     
     def set_scale_multiplier(self, multiplier: Union[float, None]):
         for block in self.transformer_blocks:
             block.set_scale_multiplier(multiplier)
 
+    def set_masks(self, masks: Tensor, min_val: float, max_val: float):
+        self.scale_min = min_val
+        self.scale_max = max_val
+        self.raw_scale_mask = masks
+
     def set_sub_idxs(self, sub_idxs: list[int]):
+        self.sub_idxs = sub_idxs
         for block in self.transformer_blocks:
             block.set_sub_idxs(sub_idxs)
 
     def forward(self, hidden_states, encoder_hidden_states=None, attention_mask=None):
-        batch, channel, height, weight = hidden_states.shape
+        batch, channel, height, width = hidden_states.shape
         residual = hidden_states
+
+        scale_mask = self.get_scale_mask(hidden_states)
 
         hidden_states = self.norm(hidden_states)
         inner_dim = hidden_states.shape[1]
         hidden_states = hidden_states.permute(0, 2, 3, 1).reshape(
-            batch, height * weight, inner_dim
+            batch, height * width, inner_dim
         )
         hidden_states = self.proj_in(hidden_states)
 
@@ -247,12 +294,13 @@ class TemporalTransformer3DModel(nn.Module):
                 encoder_hidden_states=encoder_hidden_states,
                 attention_mask=attention_mask,
                 video_length=self.video_length,
+                scale_mask=scale_mask
             )
 
         # output
         hidden_states = self.proj_out(hidden_states)
         hidden_states = (
-            hidden_states.reshape(batch, height, weight, inner_dim)
+            hidden_states.reshape(batch, height, width, inner_dim)
             .permute(0, 3, 1, 2)
             .contiguous()
         )
@@ -327,6 +375,7 @@ class TemporalTransformerBlock(nn.Module):
         encoder_hidden_states=None,
         attention_mask=None,
         video_length=None,
+        scale_mask=None
     ):
         for attention_block, norm in zip(self.attention_blocks, self.norms):
             norm_hidden_states = norm(hidden_states)
@@ -338,6 +387,7 @@ class TemporalTransformerBlock(nn.Module):
                     else None,
                     attention_mask=attention_mask,
                     video_length=video_length,
+                    scale_mask=scale_mask
                 )
                 + hidden_states
             )
@@ -406,7 +456,7 @@ class VersatileAttention(CrossAttentionMM):
         if multiplier is None or math.isclose(multiplier, 1.0):
             self.scale = None
         else:
-            self.scale = self.default_scale * multiplier
+            self.scale = multiplier
 
     def set_sub_idxs(self, sub_idxs: list[int]):
         if self.pos_encoder != None:
@@ -418,6 +468,7 @@ class VersatileAttention(CrossAttentionMM):
         encoder_hidden_states=None,
         attention_mask=None,
         video_length=None,
+        scale_mask=None,
     ):
         if self.attention_mode != "Temporal":
             raise NotImplementedError
@@ -441,6 +492,7 @@ class VersatileAttention(CrossAttentionMM):
             encoder_hidden_states,
             value=None,
             mask=attention_mask,
+            scale_mask=scale_mask,
         )
 
         hidden_states = rearrange(hidden_states, "(b d) f c -> (b f) d c", d=d)
