+++ animatediff/motion_utils.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 from typing import Union
+from einops import rearrange
 
 import torch
 import torch.nn.functional as F
@@ -9,6 +10,8 @@ import comfy.model_management as model_management
 import comfy.ops
 from comfy.cli_args import args
 from comfy.ldm.modules.attention import attention_basic, attention_pytorch, attention_split, attention_sub_quad, default
+from controlnet import broadcast_image_to
+from utils import repeat_to_batch_size
 from .motion_lora import MotionLoRAInfo
 
 # until xformers bug is fixed, do not use xformers for VersatileAttention! TODO: change this when fix is out
@@ -43,22 +46,78 @@ class CrossAttentionMM(nn.Module):
 
         self.to_out = nn.Sequential(operations.Linear(inner_dim, query_dim, dtype=dtype, device=device), nn.Dropout(dropout))
 
-    def forward(self, x, context=None, value=None, mask=None):
+    def forward(self, x, context=None, value=None, mask=None, scale_mask=None):
         q = self.to_q(x)
         context = default(context, x)
-        k = self.to_k(context)
+        k: Tensor = self.to_k(context)
         if value is not None:
             v = self.to_v(value)
             del value
         else:
             v = self.to_v(context)
 
-        # apply custom scale by multiplying k by scale factor;
-        # division by default_scale is needed to account for internal attn code multiplying by default_scale
+        # apply custom scale by multiplying k by scale factor
         if self.scale is not None:
-            k *= (self.scale / self.default_scale)
+            k *= self.scale
+        
+        # apply scale mask, if present
+        if scale_mask is not None:
+            k *= scale_mask
+
         out = optimized_attention_mm(q, k, v, self.heads, mask)
         return self.to_out(out)
+
+
+# super class to TemporalTransformer-like classes
+class TemporalTransformerGeneric:
+    def temporal_transformer_init(self, default_length: int):
+        self.video_length = default_length
+        self.full_length = default_length
+        self.scale_min = 1.0
+        self.scale_max = 1.0
+        self.raw_scale_mask: Union[Tensor, None] = None
+        self.temp_scale_mask: Union[Tensor, None] = None
+        self.sub_idxs: Union[list[int], None] = None
+
+    def reset_temp_vars(self):
+        self.temp_scale_mask = None
+
+    def get_scale_mask(self, hidden_states: Tensor) -> Union[Tensor, None]:
+        # if no raw mask, return None
+        if self.raw_scale_mask is None:
+            return None
+        # if temp mask already calculated, return it
+        if self.temp_scale_mask != None:
+            if self.sub_idxs is not None:
+                return self.temp_scale_mask[:, self.sub_idxs, :]
+            return self.temp_scale_mask
+        # otherwise, calculate temp mask
+        shape = hidden_states.shape
+        batch, channel, height, width = shape
+        mask = prepare_mask_batch(self.raw_scale_mask, shape=(self.full_length, 1, height, width))
+        mask = repeat_to_batch_size(mask, self.full_length)
+        # if mask not the same amount length as full length, make it match
+        if self.full_length != mask.shape[0]:
+            mask = broadcast_image_to(mask, self.full_length, 1)
+        # reshape mask to attention K shape (h*w, latent_count, 1)
+        batch, channel, height, width = mask.shape
+        # first, perform same operations as on hidden_states,
+        # turning (b, c, h, w) -> (b, h*w, c)
+        mask = mask.permute(0, 2, 3, 1).reshape(batch, height*width, channel)
+        # then, make it the same shape as attention's k, (h*w, b, c)
+        mask = mask.permute(1, 0, 2)
+        # make masks match the expected length of h*w
+        batched_number = shape[0] // self.video_length
+        if batched_number > 1:
+            mask = torch.cat([mask] * batched_number, dim=0)
+        # cache mask and set to proper device
+        self.temp_scale_mask = mask
+        # move temp_scale_mask to proper dtype + device
+        self.temp_scale_mask = self.temp_scale_mask.to(dtype=hidden_states.dtype, device=hidden_states.device)
+        # return subset of masks, if needed
+        if self.sub_idxs is not None:
+            return self.temp_scale_mask[:, self.sub_idxs, :]
+        return self.temp_scale_mask
 
 
 class BlockType:
@@ -91,19 +150,35 @@ class GenericMotionWrapper(nn.Module, ABC):
         return self.loras is not None
 
     @abstractmethod
-    def set_video_length(self, video_length: int):
+    def set_video_length(self, video_length: int, full_length: int):
         pass
 
     @abstractmethod
     def set_scale_multiplier(self, multiplier: Union[float, None]):
         pass
 
-    def reset_scale_multiplier(self):
-        self.set_scale_multiplier(None)
+    @abstractmethod
+    def set_masks(self, masks: Tensor, min_val: float, max_val: float):
+        pass
 
     @abstractmethod
     def set_sub_idxs(self, sub_idxs: list[int]):
         pass
+    
+    @abstractmethod
+    def reset_temp_vars(self):
+        pass
+
+    def reset_scale_multiplier(self):
+        self.set_scale_multiplier(None)
+
+    def reset_sub_idxs(self):
+        self.set_sub_idxs(None)
+
+    def reset(self):
+        self.reset_sub_idxs()
+        self.reset_scale_multiplier()
+        self.reset_temp_vars()
 
 
 class GroupNormAD(torch.nn.GroupNorm):
@@ -114,3 +189,19 @@ class GroupNormAD(torch.nn.GroupNorm):
     def forward(self, input: Tensor) -> Tensor:
         return F.group_norm(
              input, self.num_groups, self.weight, self.bias, self.eps)
+
+
+# applies min-max normalization, from:
+# https://stackoverflow.com/questions/68791508/min-max-normalization-of-a-tensor-in-pytorch
+def normalize_min_max(x: Tensor, new_min = 0.0, new_max = 1.0):
+    x_min, x_max = x.min(), x.max()
+    return (((x - x_min)/(x_max - x_min)) * (new_max - new_min)) + new_min
+
+
+# adapted from comfy/sample.py
+def prepare_mask_batch(mask: Tensor, shape: Tensor, multiplier: int=1, match_dim1=False):
+    mask = mask.clone()
+    mask = torch.nn.functional.interpolate(mask.reshape((-1, 1, mask.shape[-2], mask.shape[-1])), size=(shape[2]*multiplier, shape[3]*multiplier), mode="bilinear")
+    if match_dim1:
+        mask = torch.cat([mask] * shape[1], dim=1)
+    return mask
