+++ animatediff/nodes_extras.py
@@ -1,0 +1,69 @@
+import torch
+
+import folder_paths
+import nodes as comfy_nodes
+from comfy.model_patcher import ModelPatcher
+from comfy.sd import load_checkpoint_guess_config
+from .logger import logger
+from .model_utils import IsChangedHelper, BetaSchedules
+from .motion_module import eject_params_from_model
+
+
+class AnimateDiffUnload:
+    def __init__(self) -> None:
+        self.change = IsChangedHelper()
+
+    @classmethod
+    def INPUT_TYPES(s):
+        return {"required": {"model": ("MODEL",)}}
+
+    RETURN_TYPES = ("MODEL",)
+    CATEGORY = "Animate Diff 🎭🅐🅓/extras"
+    FUNCTION = "unload_motion_modules"
+
+    def unload_motion_modules(self, model: ModelPatcher):
+        # return model clone with ejected params
+        model = eject_params_from_model(model)
+
+        return (model,)
+
+
+class CheckpointLoaderSimpleWithNoiseSelect:
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "ckpt_name": (folder_paths.get_filename_list("checkpoints"), ),
+                "beta_schedule": (BetaSchedules.get_alias_list_with_first_element(BetaSchedules.LINEAR), )
+            },
+        }
+    RETURN_TYPES = ("MODEL", "CLIP", "VAE")
+    FUNCTION = "load_checkpoint"
+
+    CATEGORY = "Animate Diff 🎭🅐🅓/extras"
+
+    def load_checkpoint(self, ckpt_name, beta_schedule, output_vae=True, output_clip=True):
+        ckpt_path = folder_paths.get_full_path("checkpoints", ckpt_name)
+        out = load_checkpoint_guess_config(ckpt_path, output_vae=True, output_clip=True, embedding_directory=folder_paths.get_folder_paths("embeddings"))
+        # register chosen beta schedule on model - convert to beta_schedule name recognized by ComfyUI
+        out[0].model.model_sampling = BetaSchedules.to_model_sampling(beta_schedule, out[0])
+        return out
+
+
+class EmptyLatentImageLarge:
+    def __init__(self, device="cpu"):
+        self.device = device
+
+    @classmethod
+    def INPUT_TYPES(s):
+        return {"required": { "width": ("INT", {"default": 512, "min": 64, "max": comfy_nodes.MAX_RESOLUTION, "step": 8}),
+                              "height": ("INT", {"default": 512, "min": 64, "max": comfy_nodes.MAX_RESOLUTION, "step": 8}),
+                              "batch_size": ("INT", {"default": 1, "min": 1, "max": 262144})}}
+    RETURN_TYPES = ("LATENT",)
+    FUNCTION = "generate"
+
+    CATEGORY = "Animate Diff 🎭🅐🅓/extras"
+
+    def generate(self, width, height, batch_size=1):
+        latent = torch.zeros([batch_size, 4, height // 8, width // 8])
+        return ({"samples":latent}, )
